refactor: extract CLI logic from main.rs into cli module (#51)
- Created src/cli/ module structure (mod.rs, commands.rs, config.rs)
- Moved load_xxhash_config and helpers to cli/config.rs
- Moved matrix_mode, detect_mode to cli/commands.rs
- Extracted streaming encode/decode to cli/commands.rs
- Moved CLI struct and orchestration to cli/mod.rs as run()
- Reduced main.rs from 585 lines to 5-line entry point

Benefits:
- Business logic now library-accessible (not CLI-only)
- Improved testability (no CLI parsing required)
- Clear separation of I/O concerns from domain logic
- Each module has focused, single responsibility

All 229 tests pass. No breaking changes.
